Combine the overview and setup pages
